Fix remaining bugs in common and modern modules
common (already compiling clean; these are logic fixes, not compile
errors, confirmed via gradle :common:compileJava after each change):
- PerformanceOptimizer.nextEnabledAction: fix a copy-paste bug where both
  branches of the perfModeGamePriority ternary built the identical
  escalation order, making the "game priority" config setting dead.
  Game-priority mode now escalates to LOWER_RESOLUTION first.
- VoiceChatIntegration.stopCapture: close the sidecar OutputStream in a
  finally block instead of only on the success path, so a flush()
  failure no longer leaks the open file handle.
- RecordableConfig.parseHexColor: use Long.parseLong instead of
  Integer.parseInt for 8-digit AARRGGBB hex colors, matching
  parseArgbColor's existing technique. Integer.parseInt overflowed
  (NumberFormatException) for any color with the alpha byte's high bit
  set, silently discarding a validly-formatted color.
- dev/gxlg/multiversion/V.getVersion: make the lazily-detected version
  field volatile with double-checked locking, matching every other
  lazy singleton in this codebase (RecordableConfig, OpenALLoopbackCapture,
  DeferredCaptureManager, etc.) since this utility is called from
  multiple threads (render/network/capture).

modern (net.minecraft.* dependent, reviewed by reading since it can't be
compiled in this sandbox):
- screen/AudioHelpScreen.java, screen/RecordableSettingsScreen.java: an
  apparent automated find/replace renaming the input-event API type
  Click -> MouseButtonEvent had corrupted plain English UI text ("3.
  MouseButtonEvent 'Test Audio' to verify..."). Restored the intended
  "Click" wording in all affected strings.
- screen/VideoCollectionScreen.isSupportedVideo: recognize .webm/.mov/.avi
  recordings, not just .mp4/.mkv (same gallery-visibility bug fixed in
  legacy).
- screen/VideoPlayerScreen.stopFrameExtraction: close the Stream<Path>
  from Files.walk (same leak fixed in legacy/crust).
- ReplayBuffer.saveFramesToFile: drain the FFmpeg render process's merged
  stdout/stderr pipe on a daemon thread before waitFor() (same
  pipe-fill-deadlock risk fixed elsewhere in this file's own
  KillClipBuffer.encode, and in legacy's ReplayBuffer).

+++ src/common/src/main/java/dev/recordable/RecordableConfig.java
@@ -2392,7 +2392,11 @@ public final class RecordableConfig {
     private static int parseHexColor(String value, int fallback) {
         String normalized = sanitizeHexColor(value, String.format(Locale.ROOT, "#%06X", fallback));
         try {
-            return Integer.parseInt(normalized.substring(1), 16);
+            // Use Long.parseLong + cast rather than Integer.parseInt: an 8-digit
+            // AARRGGBB value with the alpha byte's high bit set (e.g. "#C8FFFFFF")
+            // exceeds Integer.MAX_VALUE and Integer.parseInt throws
+            // NumberFormatException for it, silently discarding a valid color.
+            return (int) Long.parseLong(normalized.substring(1), 16);
         } catch (Throwable ignored) {
             return fallback;
         }
